+++ sdk/containerregistry/azure-containerregistry/azure/containerregistry/__init__.py
@@ -10,10 +10,10 @@ from ._container_registry_client import ContainerRegistryClient
 from ._models import (
     ArtifactArchitecture,
     ArtifactOperatingSystem,
-    ManifestOrder,
+    ArtifactManifestOrder,
     ArtifactManifestProperties,
     RepositoryProperties,
-    TagOrder,
+    ArtifactTagOrder,
     ArtifactTagProperties,
 )
 from ._version import VERSION
@@ -24,9 +24,9 @@ __all__ = [
     "ArtifactArchitecture",
     "ArtifactOperatingSystem",
     "ContainerRegistryClient",
-    "ManifestOrder",
+    "ArtifactManifestOrder",
     "ArtifactManifestProperties",
     "RepositoryProperties",
-    "TagOrder",
+    "ArtifactTagOrder",
     "ArtifactTagProperties",
 ]